+++ shared-helpers/src/views/components/ClickableCard.module.scss
