+++ sites/public/src/components/browse/ListingCard.tsx
@@ -5,11 +5,10 @@ import {
   ListingsStatusEnum,
   ReviewOrderTypeEnum,
 } from "@bloom-housing/shared-helpers/src/types/backend-swagger"
-import { imageUrlFromListing, oneLineAddress } from "@bloom-housing/shared-helpers"
+import { imageUrlFromListing, oneLineAddress, ClickableCard } from "@bloom-housing/shared-helpers"
 import { StackedTable, t } from "@bloom-housing/ui-components"
 import { Card, Heading, Icon, Link, Message, Tag } from "@bloom-housing/ui-seeds"
 import { getListingApplicationStatus, getListingStackedTableData } from "../../lib/helpers"
-import { ClickableCard } from "../../patterns/ClickableCard"
 import { getListingTags } from "../listing/listing_sections/MainDetails"
 import styles from "./ListingCard.module.scss"
